# [KT.regroup Ops][4/N] benchmark of fbgemm op - regroup_kts